+++ frontend/src/App.tsx
@@ -2,15 +2,16 @@ import { Routes, Route } from "react-router-dom";
 import { GlobalStyle, lightTheme, darkTheme } from "styles";
 import { Credits, Main, ProfileEdit, Register, StudentCheck, ProfileRegister } from "pages";
 import { paths } from "consts";
-import { useState, useEffect } from "react";
-import { useSetRecoilState } from "recoil";
-import { scrollPosState } from "atoms";
+import { useEffect } from "react";
+import { useRecoilValue, useSetRecoilState } from "recoil";
+import { isDarkThemeState, scrollPosState } from "atoms";
 import { throttle } from "lodash";
-import styled, { ThemeProvider } from "styled-components";
+import { ThemeProvider } from "styled-components";
+import { ThemeButton } from "components";
 
 export default function App() {
   const setScrollPos = useSetRecoilState(scrollPosState);
-  const [isDarkTheme, setDarkTheme] = useState(false);
+  const isDarkTheme = useRecoilValue(isDarkThemeState);
 
   useEffect(() => {
     window.addEventListener(
@@ -20,10 +21,6 @@ export default function App() {
       }, 100)
     );
   }, []);
-
-  const changeTheme = () => {
-    setDarkTheme(!isDarkTheme);
-  };
 
   return (
     <>
@@ -38,41 +35,9 @@ export default function App() {
             <Route path="profile-register" element={<ProfileRegister />} />
           </Route>
         </Routes>
-
-        <ThemeButton onClick={changeTheme}>
-          <img src={isDarkTheme ? "/icons/sun.png" : "/icons/moon.png"} />
-        </ThemeButton>
+        <ThemeButton />
         <GlobalStyle />
       </ThemeProvider>
     </>
   );
 }
-
-const ThemeButton = styled.button`
-  position: fixed;
-  right: 20px;
-  bottom: 100px;
-
-  width: 50px;
-  height: 50px;
-
-  background-color: ${({ theme }) => theme.colors.themeBtn};
-  border-radius: 70%;
-
-  box-shadow: 1px 1px 10px 1px rgba(0, 0, 0, 0.5);
-
-  opacity: 0.3;
-
-  transition: opacity 0.3s, background-color 1s;
-
-  img {
-    width: 40px;
-    height: 40px;
-    margin-left: 5px;
-    margin-top: 3px;
-  }
-
-  &:hover {
-    opacity: 1;
-  }
-`;
