+++ frontend/src/components/index.ts
@@ -3,3 +3,4 @@ export { default as ProfileBox } from "./ProfileBox";
 export { default as TopBackground } from "./common/TopBackground";
 export { default as TopBar } from "./common/TopBar";
 export { default as RegisterPageTitle } from "./RegisterPageTitle";
+export { default as ThemeButton } from "./ThemeButton";
